refactor(view-engine): rename sessionValidation to sessionValidationCache
